migrate deps
